cast input_event.value as (int) and negatives work;  why do they declare it unsigned int if they expect negatives?
git-svn-id: https://pure-data.svn.sourceforge.net/svnroot/pure-data/trunk@568 df7eb9d6-776f-4174-a42d-0ffd8d5d5b5d

+++ externals/hcs/rawmouse.c
@@ -292,7 +292,7 @@ static int rawmouse_read(t_rawmouse *x,int fd)
       axis_num = 2;
       break;
     }
-    outlet_float (x->x_axis_out[axis_num], x->x_input_event.value);	
+    outlet_float (x->x_axis_out[axis_num], (int)x->x_input_event.value);	
   }
 
   return 1;
